requirements.txt
streamlit
writerai
requests
pillow